% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_input.R
\name{check_input}
\alias{check_input}
\title{Checks input from caller}
\usage{
check_input(caller, args_to_check)
}
\description{
Performs a sanity check on the inputs from caller. It stops execution and
outputs an error message if arguments are not in conformity with caller
method.
}
