% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_dynseg.R
\name{compute_dynseg}
\alias{compute_dynseg}
\title{Block segmentation using dynamical programming}
\usage{
compute_dynseg(
  suff_stats,
  family,
  lambda = 1,
  n,
  m,
  max_blocks = m - 1,
  pen_func = bic_loss
)
}
\arguments{
\item{suff_stats}{Sufficient statistics to perform change point analysis}

\item{family}{The name of the family used to fit the model}

\item{lambda}{Penalization constant.}

\item{n}{Number of samples}

\item{m}{Number of variables}

\item{max_blocks}{Threshold on the number of block segments to fit the model.
Set low values for this parameters if having performance issues on large
data sets.}

\item{pen_func}{A penalization function defined i integer intervals
The function signature should be pen(left_index, right_index, n, m),
where the left_index:right_index is the integer interval, n the sample
size and m the number of variables/columns.}
}
\description{
Computes the exact solution of the regularized loss optimization problem,
providing change point locations and the parameters of each blocks. Should
be called within \link[=fit_blockcpd]{fit_blockcpd}
}
