% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{compute_hausdorff}
\alias{compute_hausdorff}
\title{Hausdorff distance metric}
\usage{
compute_hausdorff(cp1, cp2)
}
\arguments{
\item{cp1}{Change point set for model 1 or true change point set.}

\item{cp2}{Change point set for model 2 or true change point set.}
}
\description{
Computes the Hausdorff distance between change point sets.
}
