% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_hierseg.R
\name{compute_hierseg}
\alias{compute_hierseg}
\title{Block segmentation using hierarchical algorithm}
\usage{
compute_hierseg(
  suff_stats,
  family,
  lambda = 1,
  n,
  m,
  pen_func = bic_loss_hs,
  max_blocks = NULL
)
}
\arguments{
\item{suff_stats}{Sufficient statistics to perform change point analysis}

\item{family}{The name of the family used to fit the model}

\item{lambda}{Penalization constant}

\item{n}{Number of samples}

\item{m}{Number of variables}

\item{pen_func}{A penalization function defined i integer intervals
The function signature should be pen(left_index, right_index, n, m),
where the left_index:right_index is the integer interval, n the sample
size and m the number of variables/columns.}
}
\description{
Uses binary splitting to obtain a greedy soltuion to the regularized loss
optimization problem. Should be called within
\link[=fit_blockcpd]{fit_blockcpd}
}
