% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{compute_jaccard}
\alias{compute_jaccard}
\title{Jaccard's Index metric}
\usage{
compute_jaccard(cp1, cp2)
}
\arguments{
\item{cp1}{Change point set for model 1 or true change point set.}

\item{cp2}{Change point set for model 2 or true change point set.}
}
\description{
Computes the Jaccard index between two change point
detection sets
}
