% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{compute_rand}
\alias{compute_rand}
\title{Rand Index Function for change point detection}
\usage{
compute_rand(cp1, cp2, m)
}
\arguments{
\item{cp1}{Change point set for model 1 or true change point set.}

\item{cp2}{Change point set for model 2 or true change point set.}

\item{m}{The size of the vector array.}
}
\description{
Computes the rand Index (non-adjusted) for the change point sets. A specific
equation for change point detection is used to make the computation faster.
Proof of correctness of the equation is given in the dissertation.
}
