% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{compute_symdiff}
\alias{compute_symdiff}
\title{Symmetric difference metric}
\usage{
compute_symdiff(cp1, cp2)
}
\arguments{
\item{cp1}{Change point set for model 1 or true change point set.}

\item{cp2}{Change point set for model 2 or true change point set.}
}
\description{
Computes the size of the symmetric difference between two change point
detection sets
}
