% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_blockcpd.R
\name{fit_blockcpd}
\alias{fit_blockcpd}
\title{Fits a blockcpd model}
\usage{
fit_blockcpd(
  data_matrix,
  method = "hierseg",
  family = "bernoulli",
  lambda = 1,
  select_lambda = FALSE,
  pen_func = bic_loss_hs,
  max_blocks = NULL,
  bootstrap = FALSE,
  bootstrap_rep = 100L,
  bootstrap_progress = FALSE
)
}
\arguments{
\item{data_matrix}{Data frame or matrix containing the data set to be
segmented. There is no verification if the entries correspond to the model
specified by the "family" argument, such as entries different than 0, 1 or NA
for the bernoulli family.}

\item{method}{The method that will be used to fit the model. The current
implemented models are:
\itemize{
\item[\link[=compute_hierseg]{hierseg}] Hierarchical segmentation, also
known as binary segmentation;
\item[\link[=compute_dynseg]{dynseg}] Dynamical programming segmentation.
}}

\item{family}{The name of the family to detect changes in parameters. The
families  currently implemented are: "normal", "bernoulli". If "normal" is
chosen, the algorithm will estimate blocks with different mean or variance.
If bernoulli is chosen, it will estimate blocks that differ by the
probability parameter.}

\item{lambda}{The penalization constant. A list of penalization constant can
be passed if the argument "select_lambda" is TRUE.}

\item{select_lambda}{A flag to decide if the BIC criterion must be used to
choose the best lambda from the list of values provided in the "lambda"
argument.}

\item{pen_func}{Regularization function used for fitting, with default as the
BIC. For user specified functions, check the template in the
\link[=toy_regularization]{regularization} regularization.rd file.}

\item{max_blocks}{An integer greater than 0 that specify the maximum number
of blocks fitted by the algorithm. It is only used if dynseg is specified in
the "method" argument.}

\item{bootstrap}{A flag to decide if bootstrap computations for the
estimation of the probability of each index being detected as a change point.
It also provides a sample of all the metrics implemented computed with
respect to the final change point set estimated.}

\item{boostrap_rep}{Number of bootstrap repetitions.}
}
\value{
The function returns a S3 object of the type blockcpd.
\itemize{
\item\link{changepoints} a list containing the set of estimated change points;
\item\link{probabilities} a list containing the estimated probability parameters
for each block;
\item\link{loss} the final loss evaluated on the entire data set for the
returned model;
\item\link{neg_loglike} The negative log likelihood of the model;
\item\link{n_cp} number of change points estimated;
\item\link{metadata} Arguments passed to fit the model;
\item\link{bootstrap_info} if bootstrap argument is true, this contains a list
of the metrics for each bootstrap sample, and contains the estimated
probability of each index being detected as a change point;
\item\link{lambda_info} If select_lambda argument is true, this contains a list
summarizing the the BIC values and the best value for lambda.
}
}
\description{
Fits a blockcpd model to find the best segmentation of the data into blocks.
Variables in each block have the same distribution and parameter,
and consecutive blocks have different parameters.
}
