% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcpd.R
\name{rcpd}
\alias{rcpd}
\title{Sampler for the CPD Block Model}
\usage{
rcpd(
  n = 100,
  m = 50,
  k = 1,
  family = "bernoulli",
  parameters = NULL,
  changepoints = NULL,
  prob_NA = 0
)
}
\arguments{
\item{n}{Sample size.}

\item{m}{Array length.}

\item{k}{Number of change points.  The number of blocks is \eqn{k + 1}. It is
overridden if changepoints is non-NULL.}

\item{family}{The family model to be sampled. Currently, bernoulli and normal
are available. Consecutive blocks have different probability parameters for
the bernoulli, and normal has different mean and variance.}

\item{parameters}{List of parameters containing \eqn{k + 1} dimensional
parameter vectors of each block. If NULL, the parameters are sampled
randomly. If specified, pass a list of probabilities for the bernoulli,
and a list of two vectors for the normal: the first is the vector of the
means, and the second the vector of the variances}

\item{changepoints}{A sorted vector of size \eqn{k} containing integers as
change point locations. The change points are between 1 and \eqn{m-1}. If
NULL, the change points are sampled uniformly in \eqn{[1, m-1]}.}

\item{prob_NA}{Probability of each entry of being NA. Default is 0.}
}
\description{
Creates a \eqn{n \times m} matrix with \eqn{k} change points. In between
change points, the random variables are i.i.d. sampled from the given family
and parameters
}
