% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_input.R
\name{check_input}
\alias{check_input}
\title{Checks input from fit_blockcpd}
\usage{
check_input(method, family, lambda)
}
\description{
Checks if the input from \item[\link[=fit_blockcpd]{fit_blockcpd}] passes
a sanity check.
}
