% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_model.R
\name{compare_model}
\alias{compare_model}
\title{Compare or evaluate model performance with respect
to other model or ground truth}
\usage{
compare_model(model1, model2, ncol = NULL)
}
\arguments{
\item{model1}{The first blockcpd object or list of sorted integers
representing the change point set.}

\item{model2}{The second blockcpd object or list of sorted integers
representing the change point set.}

\item{ncol}{The number of variables which the model was fitted on. Only needs
to be passed if both arguments are change point sets instead of a blockcpd
object.}
}
\description{
Compares or evaluates model estimated change point set against
another model or ground truth. The comparison is made using common metrics
to compare clusters. The metrics provided are
\itemize{
\item\link{"hausdorff"} Hausdorff Distance metric;
\item\link{"rand"} Rand Index ;
\item\link{"symdiff"} Symmetric difference metric;
\item\link{"jaccard"} Jaccard similarity index.
}
}
