% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.blockcpd}
\alias{plot.blockcpd}
\title{Plot for blockcpd object}
\usage{
\method{plot}{blockcpd}(blockcpd_obj, parameter = NULL, pkg = "base")
}
\arguments{
\item{blockcpd_obj}{A fitted blockcpd S3 object provided by the
\link[=fit_blockcpd]{fit_blockcpd} function.}

\item{parameter}{The parameter of the family for which to plot the blocked}

\item{pkg}{Graphical package to be used for plotting. Current values are
"base".}
}
\description{
Plots the selected parameters in a blocked fashion.
}
