% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcpd.R
\name{rcpd}
\alias{rcpd}
\title{Sampler for the CPD Block Model}
\usage{
rcpd(
  n = 100,
  m = 50,
  k = 1,
  family = "bernoulli",
  parameters = NULL,
  changepoints = NULL,
  prob_NA = 0
)
}
\arguments{
\item{n}{Sample size.}

\item{m}{Array length.}

\item{k}{Number of change points.  The number of blocks is \eqn{k + 1}. It is
overridden if changepoints is non-NULL.}

\item{family}{The family model to be sampled. The families currently
implemented are:

\itemize{
\item\link{bernoulli} Sample independent Bernoullis with probability parameter
of the block
\item\link{normal} Sample independent Normal with mean and variance specified
by the block.
\item\link{binaryMarkov} Samples a two state Markov Chain process with transition
matrix defined by the block.
\item\link{exponential} Sample independent Exponential with scale parameter
defined by the block.
\item\link{poisson} Sample independent Poisson with rate parameter defined
by the block.
}}

\item{parameters}{List of parameters containing \eqn{k + 1} dimensional
parameter vectors of each block. If NULL, the parameters are sampled
randomly.}

\item{changepoints}{A sorted vector of size \eqn{k} containing integers as
change point locations. The change points are between 1 and \eqn{m-1}. If
NULL, the change points are sampled uniformly in \eqn{[1, m-1]}.}

\item{prob_NA}{Probability of each entry of being NA. Default is 0.}
}
\description{
Creates a \eqn{n \times m} matrix with \eqn{k} change points. In between
change points, the random variables are i.i.d. sampled from the given family
and parameters
}
