% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcpd.R
\name{rcpd}
\alias{rcpd}
\title{Sampler for the CPD Block Model}
\usage{
rcpd(
  nrow = 100,
  ncol = 50,
  ncp = 1,
  family = "bernoulli",
  parameters = NULL,
  changepoints = NULL,
  prob_NA = 0
)
}
\arguments{
\item{nrow}{Number of rows, or sample size, of the data.}

\item{ncol}{Number of columns of data matrix. It is the number of variables for
each sample.}

\item{ncp}{Number of change points.  The number of blocks is \eqn{ncp + 1}.
It is overridden if changepoints is non-NULL.}

\item{family}{The family model to be sampled. The families currently
implemented are:

\itemize{
\item\link{bernoulli} Sample independent Bernoullis with probability parameter
of the block
\item\link{normal} Sample independent Normal with mean and variance specified
by the block.
\item\link{binaryMarkov} Samples a two state Markov Chain process with transition
matrix defined by the block.
\item\link{exponential} Sample independent Exponential with scale parameter
defined by the block.
\item\link{poisson} Sample independent Poisson with rate parameter defined
by the block.
}}

\item{parameters}{List of parameters containing \eqn{ncp + 1} dimensional
parameter vectors of each block. If NULL, the parameters are sampled
randomly.}

\item{changepoints}{A sorted vector of size \eqn{ncp} containing integers as
change point locations. The change points are between 1 and \eqn{ncol-1}. If
NULL, the change points are sampled uniformly in \eqn{[1, ncol-1]}.}

\item{prob_NA}{Probability of each entry of being NA. Default is 0.}
}
\description{
Creates a \eqn{nrow \times ncol} matrix with \eqn{ncp} change points.
In between change points, the random variables are i.i.d. sampled from the
given family and parameters
}
